Move NATS Streaming client to new backingServices dir
